+++ src/main/java/webserver/handler/RequestHandler.java
@@ -4,6 +4,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import webserver.controller.RequestMapper;
 import webserver.request.HttpRequest;
+import webserver.request.RequestParser;
 import webserver.response.HttpResponse;
 
 import java.io.BufferedReader;
